c updated

+++ exercises/average.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
 int main() {
-    double num1, num2, num3, average;
+    int num1, num2, num3, average;
 
     printf("Enter three numbers: ");
-    scanf("%lf %lf %lf", &num1, &num2, &num3);
+    scanf("%d %d %d", &num1, &num2, &num3);
 
     average = (num1 + num2 + num3) / 3.0;
 
-    printf("Average = %.2f\n", average);
+    printf("Average = %d\n", average);
 
     return 0;
 }
